+++ realm/realm-library/src/androidTest/kotlin/io/realm/NullMixedSetTester.kt
@@ -229,6 +229,8 @@ class NullRealmAnySetTester(
 
     override fun setters() = Unit // Not tested
 
+    override fun dynamic() = Unit // Not tested
+
     override fun addRealmChangeListener() = Unit
 
     override fun removeSetChangeListener() = Unit