+++ realm/realm-library/src/androidTest/kotlin/io/realm/RealmModelManagedSetTester.kt
@@ -16,13 +16,10 @@
 
 package io.realm
 
-import io.realm.entities.DogPrimaryKey
-import io.realm.entities.SetContainerClass
+import io.realm.entities.*
 import io.realm.kotlin.createObject
 import io.realm.rule.BlockingLooperThread
 import java.lang.IllegalArgumentException
-import io.realm.entities.SetAllTypes
-import io.realm.entities.SetAllTypesPrimaryKey
 import kotlin.reflect.KFunction1
 import kotlin.reflect.KFunction2
 import kotlin.reflect.KMutableProperty1
@@ -37,6 +34,8 @@ import kotlin.test.*
  */
 class RealmModelManagedSetTester<T : Any>(
     private val testerName: String,
+    private val setFieldName: String,
+    private val setFieldClass: Class<T>,
     private val realmAnyType: RealmAny.Type? = null,
     private val setGetter: KFunction1<SetAllTypes, RealmSet<T>>,
     private val setSetter: KFunction2<SetAllTypes, RealmSet<T>, Unit>,
@@ -81,6 +80,8 @@ class RealmModelManagedSetTester<T : Any>(
 
         this.managedTester = ManagedSetTester(
             testerName = testerName,
+            setFieldClass = setFieldClass,
+            setFieldName = setFieldName,
             realmAnyType = realmAnyType,
             setGetter = setGetter,
             setSetter = setSetter,
@@ -147,6 +148,105 @@ class RealmModelManagedSetTester<T : Any>(
     override fun toArray() = managedTester.toArray()
 
     override fun toArrayWithParameter() = managedTester.toArrayWithParameter()
+
+    override fun dynamic() {
+        if (realmAnyType != null) {
+            dynamicRealmAnyTest()
+        } else {
+            dynamicObjectTest()
+        }
+    }
+
+    private fun toDynamicRealmAny(realm: DynamicRealm, value: RealmAny): RealmAny {
+        val id = value.asRealmModel(DogPrimaryKey::class.java).id
+        return RealmAny.valueOf(realm.where(DogPrimaryKey.CLASS_NAME).equalTo(DogPrimaryKey.ID, id).findFirst()!!)
+    }
+
+    private fun dynamicRealmAnyTest() {
+        // Create a set from a immutable schema context
+        val set = initAndAssertEmptySet()
+        realm.executeTransaction {
+            set.addAll(unmanagedInitializedSet)
+        }
+
+        val dynamicRealm = DynamicRealm.getInstance(realm.configuration)
+        val dynamicObject: DynamicRealmObject = dynamicRealm.where(SetAllTypes.NAME)
+            .equalTo(AllTypes.FIELD_STRING, "unmanaged").findFirst()!!
+        val dynamicSet: RealmSet<RealmAny> = dynamicObject.getRealmSet(setFieldName, setFieldClass) as RealmSet<RealmAny>
+
+        // Access the previous set from a mutable context
+        assertEquals(3, dynamicSet.size)
+
+        // Update the set with a new value
+        dynamicRealm.executeTransaction {
+            dynamicSet.add(toDynamicRealmAny(dynamicRealm, unmanagedNotPresentValue as RealmAny))
+        }
+
+        assertEquals(4, dynamicSet.size)
+
+        // Try to replace the whole set by a new one
+        dynamicRealm.executeTransaction {
+            dynamicObject.setRealmSet(setFieldName, RealmSet<RealmAny>().apply {
+                add(toDynamicRealmAny(dynamicRealm, unmanagedNotPresentValue as RealmAny))
+            })
+        }
+
+        assertEquals(1, dynamicObject.getRealmSet(setFieldName, setFieldClass).size)
+
+        dynamicRealm.close()
+    }
+
+    private fun dynamicObjectTest() {
+        // Create a set from a immutable schema context
+        val set = initAndAssertEmptySet()
+        realm.executeTransaction {
+            set.addAll(unmanagedInitializedSet)
+        }
+
+        val dynamicRealm = DynamicRealm.getInstance(realm.configuration)
+        val dynamicObject: DynamicRealmObject =
+            dynamicRealm.where(SetAllTypes.NAME).equalTo(AllTypes.FIELD_STRING, "unmanaged").findFirst()!!
+        val dynamicSet = dynamicObject.getRealmSet(setFieldName)
+
+        // Access the previous set from a mutable context
+        set.forEach { value ->
+            if (RealmObject.isValid(value as DogPrimaryKey)) {
+                val managedObject =
+                    dynamicRealm.where(DogPrimaryKey.CLASS_NAME).equalTo(DogPrimaryKey.ID, (value as DogPrimaryKey).id)
+                        .findFirst()!!
+                assertTrue(dynamicSet.contains(managedObject))
+            }
+        }
+
+        // Update the set with a new value
+        dynamicRealm.executeTransaction {
+            val notPresentManaged = dynamicRealm.where(DogPrimaryKey.CLASS_NAME)
+                .equalTo(DogPrimaryKey.ID, (unmanagedNotPresentValue as DogPrimaryKey).id).findFirst()!!
+            dynamicSet.add(notPresentManaged)
+        }
+
+        set.plus(unmanagedNotPresentValue).forEach { value ->
+            if (RealmObject.isValid(value as DogPrimaryKey)) {
+                val managedObject =
+                    dynamicRealm.where(DogPrimaryKey.CLASS_NAME).equalTo(DogPrimaryKey.ID, (value as DogPrimaryKey).id)
+                        .findFirst()!!
+                assertTrue(dynamicSet.contains(managedObject))
+            }
+        }
+
+        // Try to replace the whole set by a new one
+        dynamicRealm.executeTransaction {
+            val notPresentManaged = dynamicRealm.where(DogPrimaryKey.CLASS_NAME)
+                .equalTo(DogPrimaryKey.ID, (unmanagedNotPresentValue as DogPrimaryKey).id).findFirst()!!
+            dynamicObject.setRealmSet(setFieldName, RealmSet<DynamicRealmObject>().apply {
+                add(notPresentManaged)
+            })
+        }
+
+        assertEquals(1, dynamicObject.get<RealmSet<T>>(setFieldName).size)
+
+        dynamicRealm.close()
+    }
 
     override fun add() {
         // Test with managed objects